changed a length type to clear compiler warning
git-svn-id: svn://orca.physics.unc.edu/Orca64@11970 fd86f90f-2c0a-422a-b972-84106d8db31b

+++ Source/Objects/Misc Objects/DatabaseSupport/mySQL/ORSqlResult.m
@@ -86,7 +86,7 @@ NSDate* MCPYear0000;
 - (id) fetchRowAsType:(MCPReturnType) aType
 {
     MYSQL_ROW		theRow;
-    uint32_t*	theLengths;
+    unsigned long*	theLengths;
     MYSQL_FIELD*	theField;
     int				i;
     id				theReturn;
